No need to generate all locales if locales-all is not built
svn path=/glibc-package/trunk/; revision=1486

+++ debian/rules.d/build.mk
@@ -110,12 +110,14 @@ $(stamp)install_%: $(stamp)check_%
 	if [ $(curpass) = libc ]; then \
 	  $(MAKE) -f debian/generate-supported.mk IN=$(DEB_SRCDIR)/localedata/SUPPORTED \
 	    OUT=debian/tmp-$(curpass)/usr/share/i18n/SUPPORTED; \
+	  if false; then \
 	  $(MAKE) -C $(DEB_BUILDDIR) -j $(NJOBS) \
 	    objdir=$(DEB_BUILDDIR) install_root=$(CURDIR)/debian/tmp-$(curpass) \
 	    localedata/install-locales; \
 	  rm -rf $(CURDIR)/debian/locales-all/usr/lib; \
 	  install -d $(CURDIR)/debian/locales-all/usr/lib; \
 	  mv $(CURDIR)/debian/tmp-libc/usr/lib/locale $(CURDIR)/debian/locales-all/usr/lib/locales-all; \
+	  fi; \
 	  (cd $(DEB_SRCDIR)/manual && texi2html -split_chapter libc.texinfo); \
 	fi
